+++ src/tracks/circular-brush/index.ts
@@ -1,0 +1,2 @@
+export { CircularBrushTrack } from './circular-brush-plot';
+export type { CircularBrushTrackOptions, CircularBrushTrackContext } from './circular-brush';
