--- src/tracks/gosling-brush/index.ts
@@ -1,2 +1,0 @@
-export { default as BrushTrack } from './brush-track';
-export { LinearBrushModel } from './linear-brush-model';
