+++ src/tracks/gosling-track/linear-brush-model.ts
