+++ tsconfig.json
@@ -29,7 +29,7 @@
       "@gosling-lang/gosling-theme": ["src/gosling-theme/index.ts"],
       "@gosling-lang/gosling-track": ["./src/tracks/gosling-track/index.ts"],
       "@gosling-lang/gosling-genomic-axis": ["./src/tracks/gosling-genomic-axis/index.ts"],
-      "@gosling-lang/gosling-brush": ["./src/tracks/gosling-brush/index.ts"],
+      "@gosling-lang/gosling-brush": ["src/tracks/circular-brush/index.ts"],
       "@gosling-lang/dummy-track": ["./src/tracks/dummy-track/index.ts"],
       "@gosling-lang/text-track": ["./src/tracks/text-track/index.ts"],
       "@pixi-manager": ["./src/pixi-manager/index.ts"],
